+++ src/index.ts
@@ -4,3 +4,4 @@ export * from './caption.js';
 export * from './timeline.js';
 export * from './scene.js';
 export * from './asset.js';
+export * from './theme-motion.js';